requirements.txt
Django==4.1.3
djangorestframework==3.14.0
autopep8=2.0.0